requirements.txt
langchain-text-splitters
langchain-community
faiss-cpu
langchain-huggingface
sentence_transformers
google 
google-generativeai
